add cli
