+++ runner/index.js
@@ -4,36 +4,9 @@ const schedule = require('node-schedule')
 const fastify = require('fastify')({
   logger: true
 })
-const config = require('./config')
-const remote = require('./remote.js')
-const local = require('./local.js')
-const provision = require('./provision')
-const persistence = require('./persistence')
 const schema = require('./schema')
 const baseUrl = 'https://some.url.com'
-
-const runCommand = (test) => {
-  if (config.stage === 'local') {
-    return local.run(test.localShell, test.name)
-  } else {
-    return remote.run(test.shell, test.name)
-  }
-}
-
-const run = async () => {
-  if (config.stage !== 'local') {
-    await provision.ensure()
-  }
-  for (let test of config.benchmarks.tests) {
-    try {
-      let result = await runCommand(test)
-      // config.log.info(result)
-      persistence.store(result)
-    } catch (e) {
-      config.log.error(e)
-    }
-  }
-}
+const run = require('./runner')
 
 // run this every day at midnight, at least
 schedule.scheduleJob('0 0 * * *', function () {
@@ -59,4 +32,5 @@ const start = async () => {
     process.exit(1)
   }
 }
+
 start()
