+++ runner/package.json
@@ -27,7 +27,8 @@
     "pino-pretty": "^2.2.3",
     "ssh-exec-plus": "^2.0.1",
     "ssh2-client": "^1.3.2",
-    "yaml": "^1.0.2"
+    "yaml": "^1.0.2",
+    "yargs": "^12.0.5"
   },
   "devDependencies": {
     "standard": "^12.0.1"
